feat: implement text input
