+++ style.css
@@ -59,6 +59,16 @@ body {
   border-radius: 10%;
 }
 
+.key:hover {
+  background-color: #e3e3e3;
+  cursor: pointer;
+}
+
+.key:active {
+  background-color: #333;
+  color: #fff;
+}
+
 .backspace_key {
   width: 120px;
 }
@@ -89,11 +99,11 @@ body {
   color: #fff;
 }
 
-.key.remove {
+/* .key.remove {
   animation: keys_animation 0.5s forwards;
-}
+} */
 
-@keyframes key_animation {
+/* @keyframes key_animation {
   0% {
     background-color: #333;
     color: #fff;
@@ -102,4 +112,4 @@ body {
     background-color: #fff;
     color: #000;
   }
-}
+} */
